pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "0.1.0"
description = ""
authors = ["ilevk <kcs93023@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_tutorial"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
psycopg2-binary = "^2.9.9"
dependency-injector = {extras = ["pydantic"], version = "^4.41.0"}
pydantic-settings = "^2.2.1"
redis = "^5.0.4"
pre-commit = "^3.7.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
